requirements.txt
fastapi
uvicorn
langchain-core
langchain-community
dotenv
sentence-transformers
pillow
streamlit
mysql.connector